+++ src/dummies/b-dummy-module-loader/b-dummy-module-loader.ts
@@ -1,0 +1,45 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+/**
+ * [[include:dummies/b-dummy-module-loader/README.md]]
+ * @packageDocumentation
+ */
+
+import iData, { component, prop, Module } from 'super/i-data/i-data';
+
+export * from 'super/i-data/i-data';
+
+@component({
+	functional: {
+		functional: true,
+		dataProvider: undefined
+	}
+})
+
+export default class bDummyModuleLoader extends iData {
+	/** @override */
+	@prop({
+		default: () => globalThis.loadFromProp === true ?
+			[
+				{
+					id: 'b-dummy-module-1',
+					load: () => import('dummies/b-dummy-module-loader/b-dummy-module-1')
+				},
+
+				{
+					id: 'b-dummy-module-2',
+					load: () => import('dummies/b-dummy-module-loader/b-dummy-module-2')
+				}
+			] :
+
+			[]
+	})
+
+	dependenciesProp!: Module[];
+}