+++ src/pages/p-v4-components-demo/index.js
@@ -47,9 +47,11 @@ package('p-v4-components-demo')
 		'b-dummy',
 		'b-dummy-text',
 		'b-dummy-async-render',
+		'b-dummy-module-loader',
 		'b-dummy-lfc',
 		'b-dummy-watch',
 		'b-dummy-sync',
+		'b-dummy-state',
 		'b-dummy-control-list',
 		'b-dummy-decorators',